Move simulator.tiles to own package
Reorganize the tiles into a `simulator.tiles` package:

```
src/aihwkit/
└── simulator
   └── tiles
       ├── __init__.py
       ├── analog.py
       ├── base.py
       ├── floating_point.py
       ├── inference.py
       └── numpy.py
```

In the same spirit as for `configs` and `nn`, convenience imports for
the actual classes are kept into `__init__`.
